pyproject.toml
[project]
name = "unirl"
version = "0.1.0"
description = "A strongly-typed, modular reinforcement learning framework built around composable interfaces."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
    "pyright>=1.1.390",
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/unirl"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
include = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
